+++ crates/vmi-utils/src/lib.rs
@@ -15,5 +15,8 @@ pub mod interceptor;
 #[cfg(feature = "ptm")]
 pub mod ptm;
 
+#[cfg(feature = "resolver")]
+pub mod resolver;
+
 mod hexdump;
 pub use self::hexdump::{Representation, hexdump};